pyproject.toml
[tool.poetry]
name = "interview_project"
version = "0.1.0"
description = ""
authors = ["Larkin Piner <larkinpiner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
attrs = "^21.2.0"
SQLAlchemy = "^1.4.26"
toml = "^0.10.2"
alembic = "^1.7.4"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
